msm: HDMI: CEC: Hardware FSM reset.
Hardware requires a FSM reset toggle under frame
write/read scenarios. This enables every transaction/feature
to have a fresh start. CEC CTS requires this patch to pass a subset
of test cases in sections 8, 9. Hardware and VI teams have confirmed
this as a necessity for compliance tests to pass.

Change-Id: Id79795dd4c17c3a75337a4c3818219ba6d87ff25
Signed-off-by: Manoj Rao <manojraj@codeaurora.org>